consumer/requirements.txt
confluent-kafka
snowflake-connector-python
